requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clize==4.1.1
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.15.1
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
od==2.0.2
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.23.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sympy==1.12.1
thop==0.1.1.post2209072238
tifffile==2024.5.22
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.31
ultralytics-thop==0.2.8
urllib3==2.2.1
Werkzeug==3.0.3
